Started working on multiline functions. Like this:
```
10 DEF ADDTWO(A#, B#)
20 C# = A# + B#
30 RETURN C#
40 PRINT ADDTWO(3, 5)
```

Expectation here is that you print `8`. Unfortunately multiline function calls are a
bastardization of userFunction and GOSUB under the covers. Basically if a DEF is missing an =
then you get a multiline construct in the BasicFunction with a lineno. Processing continues
via waitingForCommand until a RETURN line is found, at which point parsing resumes. A call
to a multiline function (where expression is nil but a line number is stored) gets a new
environment set up just like a regular user function. But instead of evaluating an expression
and immediately returning it, the behavior acts like GOSUB, and the userFunction naively
returns a static true value immediately. This results in the user function executing completely,
and the return value does get calculated and returned, but by that time the original function call
has actually exited so the return value is not placed where it belongs. The actual execution flow
is more like:

1. Define function
1.1 Function Body
1.2 Function Return
2. Assign Function Return to Variable
2.1 Line in 2 completes
2.2 Function executes
2.3 Function returns
2.4 Function return value is lost

I am not sure how to fix this. We need to essentially pause the processing of the current line,
execute the function, and then return to finish executing the rest of the line. I don't see a good
way to do that currently. Maybe if I made the environment itself capable of executing code - so the
line scanning, parsing and executing actually moved inside of the environment. Maybe then I could
have the new environment for the user function pick up execution on the function call, then pass that
back out to the parent environment when it returns, and then the parent environment could continue
executing.

That feels like a pretty radical change. Not sure how ready I am to tackle that right now.